requirements.txt
anthropic>=0.40.0
hubspot-api-client>=8.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
